break linewise into 2 files for better parallelization
